changed yang_id and redistribution_id to u32

+++ src/interfaces/IShrine.cairo
@@ -10,7 +10,7 @@ trait IShrine {
     fn get_trove(trove_id: u64) -> Trove;
     fn get_yin(user: ContractAddress) -> Wad;
     fn get_yang_total(yang: ContractAddress) -> Wad;
-    fn get_yangs_count() -> u64;
+    fn get_yangs_count() -> u32;
     fn get_deposit(yang: ContractAddress, trove_id: u64) -> Wad;
     fn get_total_debt() -> Wad;
     fn get_total_yin() -> Wad;
@@ -19,8 +19,8 @@ trait IShrine {
     fn get_multiplier(interval: u64) -> (Ray, Ray);
     fn get_yang_threshold(yang: ContractAddress) -> Ray;
     fn get_redistributions_count() -> u64;
-    fn get_trove_redistribution_id(trove_id: u64) -> u64;
-    fn get_redistributed_unit_debt_for_yang(yang: ContractAddress, redistribution_id: u64) -> Wad;
+    fn get_trove_redistribution_id(trove_id: u64) -> u32;
+    fn get_redistributed_unit_debt_for_yang(yang: ContractAddress, redistribution_id: u32) -> Wad;
     fn get_live() -> bool;
     // external
     fn add_yang(
